refactor: move PostHog queries into lib/queries/posthog
