+++ src/app/(main)/dashboard/components/visits-chart.client.tsx
@@ -2,7 +2,7 @@
 
 import { Card } from "@heroui/react";
 import { ChartTooltip, LineChart } from "@heroui-pro/react";
-import type { Visit } from "@/lib/posthog";
+import type { Visit } from "@/lib/queries/posthog";
 
 interface VisitsChartClientProps {
   data: Visit[];
